[finishes #161870458] implements unit tests for POST auth/signup endpoint
