+++ .travis.yml
@@ -2,6 +2,6 @@ language: node_js
 node_js:
   - "node"
 after_success:
-  - npm install istanbul
+  - npm install nyc
   - npm install coveralls
   - npm run coveralls
